requirements.txt
numpy>=1.24
opencv-python>=4.8
pillow>=10.0
scikit-image>=0.23

scikit-learn>=1.4
joblib>=1.3

tqdm>=4.66

fastapi>=0.111
uvicorn[standard]>=0.30
python-multipart>=0.0.9